+++ tests/Feature/AnalyseCommandTest.php
@@ -15,7 +15,7 @@ it('fails an non-existing package-directory', function () {
 });
 
 it('prints number of analysis steps', function () {
-    $this->artisan('analyse '.$this->temporaryDirectory)->expectsOutputToContain('Ran 16 analysis steps');
+    $this->artisan('analyse '.$this->temporaryDirectory)->expectsOutputToContain('Ran 17 analysis steps');
 });
 
 it('has success emoji for successful analyse step', function () {
@@ -58,6 +58,7 @@ it('succeeds for failureless analysis', function () {
 
     if (! file_exists($this->temporaryDirectory.DIRECTORY_SEPARATOR.'vendor'.DIRECTORY_SEPARATOR.'bin')) {
         \mkdir($this->temporaryDirectory.DIRECTORY_SEPARATOR.'vendor'.DIRECTORY_SEPARATOR.'bin', 0777, true);
+        \touch($this->temporaryDirectory.DIRECTORY_SEPARATOR.'vendor'.DIRECTORY_SEPARATOR.'bin'.DIRECTORY_SEPARATOR.'phpstan');
         \touch($this->temporaryDirectory.DIRECTORY_SEPARATOR.'vendor'.DIRECTORY_SEPARATOR.'bin'.DIRECTORY_SEPARATOR.'pint');
     }
     \touch($this->temporaryDirectory.DIRECTORY_SEPARATOR.'box.json');